requirements.txt
backports.ssl-match-hostname==3.4.0.2
conda===unknown
conda-env==0.1.0
Cython==0.22
gnureadline==6.3.3
ipython==3.0.0
Jinja2==2.7.3
keyring==5.3
MarkupSafe==0.23
matplotlib==1.4.3
mysql-connector-repackaged==0.3.1
nose==1.3.4
numpy==1.9.2
pandas==0.15.2
PIL==1.1.7
pip-tools==0.3.5
pycosat==0.6.0
-e git://github.com/lisa-lab/pylearn2.git@e9338af5050b7dda2fa87204d4eee89a3ed8fbdb#egg=pylearn2-master
pyparsing==2.0.3
python-dateutil==2.4.0
pytz==2014.10
PyYAML==3.11
pyzmq==14.5.0
requests==2.5.3
scikit-learn==0.15.2
scipy==0.15.1
six==1.9.0
Theano==0.6.0
tornado==4.1
